Доработка модулей отрисовки полноэкранного изображения
